moved github related code in github packages
